feat: retriever function, weight provider, case base adapter, case base system, etc.
