+++ casebased/case_base_adapter.py
@@ -1,0 +1,8 @@
+from typing import Protocol
+from casebased.components.vocabulary import Case
+
+
+class CaseBaseAdapter(Protocol):
+    def get_all_cases(self) -> list[Case]: ...
+    def create_case(self, case: Case): ...
+    def change_utility(self, case: Case, utility: int): ...
