--- casebased/components/constants.py
@@ -1,14 +1,0 @@
-from enum import Enum
-
-
-class CBTypes(Enum):
-    DF = "df"
-    KD_TREE = "kd_tree"
-
-
-class Extensions(Enum):
-    SQL = ".sql"
-    CSV = ".csv"
-    EXCEL = ".xlsx"
-    JSON = ".json"
-    XML = ".xml"
