+++ casebased/components/similarity_measure/__init__.py
@@ -1,4 +1,5 @@
 from .schema import SimilaritySchema
 from .types import SimilarityFunction
+from .weight import WeightProvider
 
-__all__ = ["SimilaritySchema", "SimilarityFunction"]
+__all__ = ["SimilaritySchema", "SimilarityFunction", "WeightProvider"]
